Rename Join Request Package Name
