+++ src/main/java/com/kevin/prayerappservice/join/JoinRequestRepository.java
@@ -1,4 +1,4 @@
-package com.kevin.prayerappservice.request;
+package com.kevin.prayerappservice.join;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
