+++ src/test/java/com/kevin/prayerappservice/service/PrayerGroupServiceTests.java
@@ -18,14 +18,13 @@ import com.kevin.prayerappservice.group.entities.PrayerGroup;
 import com.kevin.prayerappservice.group.entities.PrayerGroupUser;
 import com.kevin.prayerappservice.group.mappers.PrayerGroupMapper;
 import com.kevin.prayerappservice.group.models.*;
-import com.kevin.prayerappservice.request.JoinRequest;
-import com.kevin.prayerappservice.request.JoinRequestRepository;
+import com.kevin.prayerappservice.join.JoinRequest;
+import com.kevin.prayerappservice.join.JoinRequestRepository;
 import com.kevin.prayerappservice.user.UserRepository;
 import com.kevin.prayerappservice.user.entities.Role;
 import com.kevin.prayerappservice.user.entities.User;
 import jakarta.persistence.EntityManager;
 import jakarta.transaction.Transactional;
-import org.assertj.core.api.Assert;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
